refactor project structure - clean up code
