+++ pyproject.toml
@@ -52,12 +52,13 @@ ci = [
 [project.urls]
 Homepage = "https://github.com/instituteofcancerresearch/SOPRANO"
 "Bug Tracker" = "https://github.com/instituteofcancerresearch/SOPRANO/issues"
-Discussions = "https://github.com/bkmarzouk/symflation/discussions"
+Discussions = "https://github.com/instituteofcancerresearch/SOPRANO/discussions"
 
 [project.scripts]
-RUN_SOPRANO = "SOPRANO.run_local_ssb_selection:main"
-GET_GENOMES = "SOPRANO.run_local_ssb_selection:download_genome"
-RUN_SOPRANO_APP = "SOPRANO.run_local_ssb_selection:local_st_app"
+soprano-run = "SOPRANO.run:run_cli"
+soprano-app = "SOPRANO.run:run_app"
+soprano-get-genome = "SOPRANO.run:download_genome"
+soprano-link-vep = "SOPRANO.run:link_vep_cache"
 
 [tool.hatch]
 version.source = "vcs"
