--- src/SOPRANO/data/homo_sapiens/106_GRCh37/Homo_sapiens.GRCh37.dna.primary_assembly.fai
@@ -1,1 +1,0 @@
-/home/kareem/.vep/homo_sapiens/106_GRCh37/Homo_sapiens.GRCh37.dna.primary_assembly.fai
