+++ src/SOPRANO/shell_utils/setup_env.sh
