+++ src/SOPRANO/utils/vep_utils.py
@@ -2,8 +2,9 @@ import argparse
 import pathlib
 from typing import List, Tuple
 
-from SOPRANO.misc_utils import Directories
-from SOPRANO.sh_utils.subprocess_pipes import pipe
+from SOPRANO.utils import sh_utils
+from SOPRANO.utils.misc_utils import Directories
+from SOPRANO.utils.sh_utils import pipe
 
 
 def _get_src_dst_link_pairs(vep_cache: pathlib.Path):
@@ -138,7 +139,42 @@ def _link_vep_cache_parser():
     return src_cache
 
 
-def link_vep_cache():
-    src_cache = _link_vep_cache_parser()
-    src_dst_links = _get_src_dst_link_pairs(src_cache)
-    _link_src_dst_pairs(src_dst_links)
+def annotate_homo_sapiens(
+    input_file: pathlib.Path,
+    assembly="GRCh38",
+    output_file: pathlib.Path | None = None,
+    overwrite=False,
+    fasta_file=Directories.data("ensemble_transcriptID.translated.fasta"),
+):
+    # TODO: Complete and test ! This is place holding...
+
+    if output_file is None:
+        output_file = input_file.with_suffix(".anno")
+
+    sh_utils.pipe(
+        [
+            "vep",
+            "-i",
+            input_file.as_posix(),
+            "-o",
+            output_file.as_posix(),
+            "--cache",
+            "--dir_cache",
+            Directories.data().as_posix(),
+            "--assembly",
+            assembly,
+            "--all_refseq",
+            "--pick",
+            "--symbol",
+            "--no_stats",
+            "--force_overwrite",  # TODO: Fix optional
+            "--fasta",
+            fasta_file.as_posix(),
+        ]
+    )
+
+
+if __name__ == "__main__":
+    # Check annotation procedure on vcf input
+    test_input = Directories.examples("homo_sapiens_GRCh38.vcf")
+    annotate_homo_sapiens(test_input)
