+++ tests/test_configuration/test_setup_conda.py
@@ -1,23 +1,29 @@
 import pytest
 
-from SOPRANO.env_utils import config_conda_env
+from SOPRANO.utils.env_utils import (
+    _has_conda,
+    _has_mamba,
+    bedtools_installed,
+    running_soprano_env,
+    vep_installed,
+)
 
 
 @pytest.mark.dependency(name="has_conda")
 def test__has_conda():
-    assert config_conda_env._has_conda() or config_conda_env._has_mamba()
+    assert _has_conda() or _has_mamba()
 
 
 @pytest.mark.dependency(depends=["has_conda"])
 def test_running_soprano_env():
-    assert config_conda_env.running_soprano_env()
+    assert running_soprano_env()
 
 
 @pytest.mark.dependency(depends=["has_conda"])
 def test_vep_installed():
-    assert config_conda_env.vep_installed()
+    assert vep_installed()
 
 
 @pytest.mark.dependency(depends=["has_conda"])
 def test_bedtools_installed():
-    assert config_conda_env.bedtools_installed()
+    assert bedtools_installed()
