+++ tests/test_units/test_dnds.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-import SOPRANO.dnds
+import SOPRANO.core.dnds
 
 
 def test__preprocess_dfs(tcga_05_4396_ssb192_cfg):
@@ -11,7 +11,7 @@ def test__preprocess_dfs(tcga_05_4396_ssb192_cfg):
         sites_intra_path,
     ) = tcga_05_4396_ssb192_cfg
 
-    merged, sites_extra, sites_intra = SOPRANO.dnds._preprocess_dfs(paths)
+    merged, sites_extra, sites_intra = SOPRANO.core.dnds._preprocess_dfs(paths)
 
     assert merged.head().equals(pd.read_csv(merged_path, delimiter="\t"))
     assert sites_extra.head().equals(
@@ -53,7 +53,7 @@ def test__compute_mutation_counts():
         ],
     )
 
-    computed_series = SOPRANO.dnds._compute_mutation_counts(mock_df)
+    computed_series = SOPRANO.core.dnds._compute_mutation_counts(mock_df)
 
     assert computed_series.equals(expected_series)
 
@@ -67,7 +67,7 @@ def test__define_variables():
         {"m_1": 1, "m_2": 2, "x_1": 3, "x_2": 4, "y_1": 5, "y_2": 6}
     )
 
-    computed_series = SOPRANO.dnds._define_variables(
+    computed_series = SOPRANO.core.dnds._define_variables(
         mock_series, mock_df_1, mock_df_2
     )
 
@@ -87,7 +87,8 @@ def test__rescale_intron_by_synonymous():
 
     expected_value = 1.0
     assert (
-        SOPRANO.dnds._rescale_intron_by_synonymous(mock_vars) == expected_value
+        SOPRANO.core.dnds._rescale_intron_by_synonymous(mock_vars)
+        == expected_value
     )
 
 
@@ -105,8 +106,8 @@ def test__compute_kaks_intra_extra():
         }
     )
 
-    assert SOPRANO.dnds._compute_kaks_extra(mock_vars) == 1
-    assert SOPRANO.dnds._compute_kaks_intra(mock_vars) == 4
+    assert SOPRANO.core.dnds._compute_kaks_extra(mock_vars) == 1
+    assert SOPRANO.core.dnds._compute_kaks_intra(mock_vars) == 4
 
 
 def test__compute_kaks_intron():
@@ -124,7 +125,7 @@ def test__compute_kaks_intron():
         }
     )
 
-    assert SOPRANO.dnds._compute_kaks_intron(mock_vars) == 1.0
+    assert SOPRANO.core.dnds._compute_kaks_intron(mock_vars) == 1.0
 
 
 def test__compute_conf_interval():
