+++ tests/test_units/test_run_local_ssb_selection.py
@@ -2,16 +2,16 @@ import pathlib
 
 import pytest
 
-from SOPRANO import run_local_ssb_selection as local_ssb
+import SOPRANO.utils.misc_utils
 
 
 def test_check_path():
     # Path is None checks
-    local_ssb.check_path(None, optional=True)
+    SOPRANO.utils.misc_utils.check_cli_path(None, optional=True)
     with pytest.raises(Exception):
-        local_ssb.check_path(None, optional=False)
+        SOPRANO.utils.misc_utils.check_cli_path(None, optional=False)
 
     # Path is not None checks
-    local_ssb.check_path(pathlib.Path.cwd())
+    SOPRANO.utils.misc_utils.check_cli_path(pathlib.Path.cwd())
     with pytest.raises(Exception):
-        local_ssb.check_path(pathlib.Path.cwd() / "foo")
+        SOPRANO.utils.misc_utils.check_cli_path(pathlib.Path.cwd() / "foo")
